RB->rb, moved noisemodels to rb
