+++ src/qibocal/calibrations/niGSC/simulfilteredrb.py
@@ -11,7 +11,7 @@ from plotly.graph_objects import Figure
 from qibo.models import Circuit
 from qibo.noise import NoiseModel
 
-import qibocal.protocols.characterization.RB.fitting as fitting_methods
+import qibocal.protocols.characterization.rb.fitting as fitting_methods
 from qibocal.calibrations.niGSC.basics.circuitfactory import SingleCliffordsFactory
 from qibocal.calibrations.niGSC.basics.experiment import Experiment
 from qibocal.calibrations.niGSC.basics.plot import Report, scatter_fit_fig
