+++ src/qibocal/calibrations/niGSC/standardrb.py
@@ -14,7 +14,7 @@ from qibo import gates
 from qibo.models import Circuit
 from qibo.noise import NoiseModel
 
-import qibocal.protocols.characterization.RB.fitting as fitting_methods
+import qibocal.protocols.characterization.rb.fitting as fitting_methods
 from qibocal.calibrations.niGSC.basics.circuitfactory import SingleCliffordsFactory
 from qibocal.calibrations.niGSC.basics.experiment import Experiment
 from qibocal.calibrations.niGSC.basics.plot import Report, scatter_fit_fig
