+++ src/qibocal/protocols/characterization/__init__.py
@@ -15,7 +15,7 @@ from .qubit_spectroscopy import qubit_spectroscopy
 from .rabi.amplitude import rabi_amplitude
 from .rabi.length import rabi_length
 from .ramsey import ramsey
-from .RB.standard_rb import standard_rb
+from .rb.standard_rb import standard_rb
 from .resonator_punchout import resonator_punchout
 from .resonator_punchout_attenuation import resonator_punchout_attenuation
 from .resonator_spectroscopy import resonator_spectroscopy
