+++ tests/test_niGSC_experiment.py
@@ -7,7 +7,7 @@ from qibo.models import Circuit
 
 from qibocal.calibrations.niGSC.basics.circuitfactory import Qibo1qGatesFactory
 from qibocal.calibrations.niGSC.basics.experiment import *
-from qibocal.calibrations.niGSC.basics.noisemodels import PauliErrorOnX
+from qibocal.protocols.characterization.rb.noisemodels import PauliErrorOnX
 
 
 @pytest.fixture
